requirements.txt
fastapi[all]
keras==2.9.0
numpy==1.23.1
opencv_python==4.4.0.46
Pillow==9.2.0
tensorflow==2.9.1
gunicorn
h5py==2.10.0
gdown
